feat: Add new settings tabs (#342)
* add download settings tab

* add customization settings tab

* remove settings from general settings tab

* add routes to new tabs

* remove setting sufix

* move sections into folder

* move 'Reset Settings' button to settings sidebar
